core/serializers.py
from core.models import Task
from rest_framework import serializers


class TaskSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Task
        fields = ['url', 'title', 'description', 'due_date', 'created', 'reporter', 'status']
